Convert dynamite Output to draws_df Format

Converts the output from a dynamite() call to a draws_df format of the posterior package, enabling the use of diagnostics and plotting methods of posterior and bayesplot packages. Note that this function returns variables in a wide format, whereas as.data.frame() uses the long format.

Usage

# S3 method for dynamitefit
as_draws_df(x, parameters = NULL, responses = NULL, types = NULL, ...)

# S3 method for dynamitefit
as_draws(x, parameters = NULL, responses = NULL, types = NULL, ...)

Arguments

x

[dynamitefit]
The model fit object.

parameters

[character()]
Parameter(s) for which the samples should be extracted. Possible options can be found with function get_parameter_names(). Default is all parameters of specific type for all responses.

responses

[character()]
Response(s) for which the samples should be extracted. Possible options are elements of unique(x$priors$response), and the default is this entire vector. Ignored if the argument parameters is supplied.

types

[character()]
Type(s) of the parameters for which the samples should be extracted. See details of possible values. Default is all values listed in details except spline coefficients omega, omega_alpha, and omega_psi. See also get_parameter_types(). Ignored if the argument parameters is supplied.

...

Ignored.

Value

A draws_df object.

A draws_df object.

Details

You can use the arguments parameters, responses and types to extract only a subset of the model parameters (i.e., only certain types of parameters related to a certain response variable).

See potential values for the types argument in as.data.frame.dynamitefit() and get_parameter_names() for potential values for parameters argument.

See also

Examples

data.table::setDTthreads(1) # For CRAN
as_draws(gaussian_example_fit, types = c("sigma", "beta"))
#> # A draws_df: 100 iterations, 2 chains, and 2 variables
#>    beta_y_z sigma_y
#> 1         2    0.20
#> 2         2    0.19
#> 3         2    0.20
#> 4         2    0.20
#> 5         2    0.19
#> 6         2    0.20
#> 7         2    0.20
#> 8         2    0.19
#> 9         2    0.20
#> 10        2    0.19
#> # ... with 190 more draws
#> # ... hidden reserved variables {'.chain', '.iteration', '.draw'}

# Compute MCMC diagnostics using the posterior package
posterior::summarise_draws(as_draws(gaussian_example_fit))
#> # A tibble: 143 × 10
#>    variable      mean median     sd    mad      q5   q95  rhat ess_bulk ess_tail
#>    <chr>        <dbl>  <dbl>  <dbl>  <dbl>   <dbl> <dbl> <dbl>    <dbl>    <dbl>
#>  1 alpha_y[2]  0.0590 0.0606 0.0293 0.0280 0.00900 0.106 0.995     235.    185. 
#>  2 alpha_y[3]  0.0931 0.0928 0.0434 0.0475 0.0285  0.161 1.00      172.    155. 
#>  3 alpha_y[4]  0.164  0.161  0.0367 0.0393 0.110   0.223 0.995     236.    178. 
#>  4 alpha_y[5]  0.259  0.259  0.0378 0.0408 0.196   0.322 1.01      262.    231. 
#>  5 alpha_y[6]  0.298  0.295  0.0351 0.0320 0.244   0.358 1.01      257.    226. 
#>  6 alpha_y[7]  0.328  0.326  0.0356 0.0322 0.270   0.386 1.00      186.    148. 
#>  7 alpha_y[8]  0.419  0.418  0.0340 0.0339 0.362   0.473 0.993     175.    228. 
#>  8 alpha_y[9]  0.455  0.451  0.0411 0.0470 0.388   0.525 0.996     188.    162. 
#>  9 alpha_y[10] 0.410  0.416  0.0445 0.0452 0.335   0.478 0.998     195.     53.8
#> 10 alpha_y[11] 0.400  0.403  0.0401 0.0390 0.339   0.469 1.00      198.    186. 
#> # ℹ 133 more rows