Get Parameter Names of the dynamite Model

Extracts all parameter names of used in the dynamitefit object.

Usage

get_parameter_names(x, types = NULL, ...)

# S3 method for class 'dynamitefit'
get_parameter_names(x, types = NULL, ...)

Arguments

x

[dynamitefit]
The model fit object.

types

[character()]
Extract only names of parameter of a certain type. See get_parameter_types().

...

Ignored.

Value

A character vector with parameter names of the input model.

Details

The naming of parameters generally follows style where the name starts with the parameter type (e.g. beta for time-invariant regression coefficient), followed by underscore and the name of the response variable, and in case of time-invariant, time-varying or random effect, the name of the predictor. An exception to this is spline coefficients omega, which also contain the number denoting the knot number.

See also

Examples

data.table::setDTthreads(1) # For CRAN
get_parameter_names(multichannel_example_fit)
#>  [1] "alpha_g"                 "beta_g_g_lag1"          
#>  [3] "beta_g_logp_lag1"        "sigma_g"                
#>  [5] "alpha_p"                 "beta_p_g_lag1"          
#>  [7] "beta_p_logp_lag1"        "beta_p_b_lag1"          
#>  [9] "alpha_b"                 "beta_b_g_lag1"          
#> [11] "beta_b_logp_lag1"        "beta_b_b_lag1"          
#> [13] "beta_b_b_lag1:logp_lag1" "beta_b_b_lag1:g_lag1"