Credible Intervals for Dynamite Model Parameters

Extracts credible intervals from dynamitefit object.

Usage

# S3 method for dynamitefit
confint(object, parm, level = 0.95, ...)

Arguments

object

[dynamitefit]
The model fit object.

parm

Ignored.

level

[numeric(1)]
Credible interval width.

...

Ignored.

Value

The rows of the resulting matrix will be named using the following logic: {parameter}_{time}_{category}_{group} where parameter is the name of the parameter, time is the time index of the parameter, category specifies the level of the response the parameter is related to if the response is categorical, and group determines which group of observations the parameter is related to in the case of random effects and loadings. Non-applicable fields in the this syntax are set to NA.

See also

Examples

confint(gaussian_example_fit, level = 0.9)
#>                                     5%          95%
#> alpha_y_1_NA_NA                     NA           NA
#> alpha_y_2_NA_NA            0.009041583  0.106502996
#> alpha_y_3_NA_NA            0.027593526  0.152159086
#> alpha_y_4_NA_NA            0.117959985  0.214737142
#> alpha_y_5_NA_NA            0.205949658  0.320061078
#> alpha_y_6_NA_NA            0.249603650  0.367018776
#> alpha_y_7_NA_NA            0.263494405  0.398843005
#> alpha_y_8_NA_NA            0.369282376  0.478092387
#> alpha_y_9_NA_NA            0.403075058  0.518357900
#> alpha_y_10_NA_NA           0.343415138  0.485008011
#> alpha_y_11_NA_NA           0.334338928  0.480202744
#> alpha_y_12_NA_NA           0.380677051  0.514803263
#> alpha_y_13_NA_NA           0.430331672  0.563230613
#> alpha_y_14_NA_NA           0.442932873  0.575931369
#> alpha_y_15_NA_NA           0.421684658  0.562559643
#> alpha_y_16_NA_NA           0.409607902  0.540694349
#> alpha_y_17_NA_NA           0.402386671  0.535555483
#> alpha_y_18_NA_NA           0.397191833  0.520395675
#> alpha_y_19_NA_NA           0.376042442  0.496462319
#> alpha_y_20_NA_NA           0.318752687  0.448958538
#> alpha_y_21_NA_NA           0.270460335  0.382840707
#> alpha_y_22_NA_NA           0.151063070  0.277742695
#> alpha_y_23_NA_NA          -0.006201836  0.129634941
#> alpha_y_24_NA_NA          -0.064458970  0.050520999
#> alpha_y_25_NA_NA          -0.028292469  0.074898081
#> alpha_y_26_NA_NA           0.004016480  0.102895892
#> alpha_y_27_NA_NA           0.012165317  0.126920842
#> alpha_y_28_NA_NA           0.030559373  0.160352527
#> alpha_y_29_NA_NA           0.162138948  0.296154539
#> alpha_y_30_NA_NA           0.661918054  0.810273238
#> beta_y_z_NA_NA_NA          1.949143583  1.985605135
#> delta_y_x_1_NA_NA                   NA           NA
#> delta_y_x_2_NA_NA         -0.208087918 -0.116002239
#> delta_y_x_3_NA_NA         -0.957353352 -0.852144437
#> delta_y_x_4_NA_NA         -0.918800835 -0.837168655
#> delta_y_x_5_NA_NA         -0.949590962 -0.877293166
#> delta_y_x_6_NA_NA         -1.093562360 -1.029874896
#> delta_y_x_7_NA_NA         -1.182421283 -1.108860109
#> delta_y_x_8_NA_NA         -1.035910851 -0.964048001
#> delta_y_x_9_NA_NA         -0.944141621 -0.862357647
#> delta_y_x_10_NA_NA        -1.040841153 -0.969944152
#> delta_y_x_11_NA_NA        -1.229449385 -1.162254853
#> delta_y_x_12_NA_NA        -1.403466025 -1.327049457
#> delta_y_x_13_NA_NA        -1.484533403 -1.421469311
#> delta_y_x_14_NA_NA        -1.486187787 -1.419338833
#> delta_y_x_15_NA_NA        -1.447982700 -1.377196934
#> delta_y_x_16_NA_NA        -1.453041642 -1.389909656
#> delta_y_x_17_NA_NA        -1.465093296 -1.384485305
#> delta_y_x_18_NA_NA        -1.402901662 -1.334600760
#> delta_y_x_19_NA_NA        -1.388381271 -1.316709074
#> delta_y_x_20_NA_NA        -1.390764853 -1.315209942
#> delta_y_x_21_NA_NA        -1.304656010 -1.241852534
#> delta_y_x_22_NA_NA        -1.255162342 -1.174192249
#> delta_y_x_23_NA_NA        -1.274475513 -1.202515193
#> delta_y_x_24_NA_NA        -1.240417726 -1.170042489
#> delta_y_x_25_NA_NA        -1.110128108 -1.037115396
#> delta_y_x_26_NA_NA        -0.928782086 -0.863832879
#> delta_y_x_27_NA_NA        -0.768412017 -0.702999912
#> delta_y_x_28_NA_NA        -0.679524594 -0.595780466
#> delta_y_x_29_NA_NA        -0.683806394 -0.578844024
#> delta_y_x_30_NA_NA        -0.263818675 -0.171071847
#> delta_y_y_lag1_1_NA_NA              NA           NA
#> delta_y_y_lag1_2_NA_NA     0.024627951  0.111798015
#> delta_y_y_lag1_3_NA_NA     0.133793988  0.202459003
#> delta_y_y_lag1_4_NA_NA     0.137933416  0.183545130
#> delta_y_y_lag1_5_NA_NA     0.138083089  0.185264719
#> delta_y_y_lag1_6_NA_NA     0.157251315  0.203551999
#> delta_y_y_lag1_7_NA_NA     0.183939696  0.232660648
#> delta_y_y_lag1_8_NA_NA     0.214266963  0.253657075
#> delta_y_y_lag1_9_NA_NA     0.261324444  0.306892085
#> delta_y_y_lag1_10_NA_NA    0.341475782  0.390989818
#> delta_y_y_lag1_11_NA_NA    0.434802124  0.479978599
#> delta_y_y_lag1_12_NA_NA    0.503921996  0.542387291
#> delta_y_y_lag1_13_NA_NA    0.510228428  0.542343646
#> delta_y_y_lag1_14_NA_NA    0.473481211  0.502247439
#> delta_y_y_lag1_15_NA_NA    0.418639797  0.452715885
#> delta_y_y_lag1_16_NA_NA    0.386455251  0.416558793
#> delta_y_y_lag1_17_NA_NA    0.376777233  0.409847589
#> delta_y_y_lag1_18_NA_NA    0.378001852  0.407814562
#> delta_y_y_lag1_19_NA_NA    0.365540882  0.399292366
#> delta_y_y_lag1_20_NA_NA    0.339419677  0.376609591
#> delta_y_y_lag1_21_NA_NA    0.310056756  0.341784961
#> delta_y_y_lag1_22_NA_NA    0.270481738  0.308962502
#> delta_y_y_lag1_23_NA_NA    0.232202290  0.272046765
#> delta_y_y_lag1_24_NA_NA    0.220563147  0.257280676
#> delta_y_y_lag1_25_NA_NA    0.213196677  0.257266677
#> delta_y_y_lag1_26_NA_NA    0.179164263  0.219851826
#> delta_y_y_lag1_27_NA_NA    0.107712683  0.158807340
#> delta_y_y_lag1_28_NA_NA    0.015904285  0.074734631
#> delta_y_y_lag1_29_NA_NA   -0.070542126 -0.003330005
#> delta_y_y_lag1_30_NA_NA    0.010001321  0.095241084
#> tau_y_x_NA_NA_NA           0.258773787  0.493718945
#> tau_y_y_lag1_NA_NA_NA      0.075351755  0.138808753
#> tau_alpha_y_NA_NA_NA       0.146597489  0.284548450
#> sigma_nu_y_alpha_NA_NA_NA  0.075670577  0.114481434
#> sigma_y_NA_NA_NA           0.191634851  0.205384386
#> nu_y_alpha_NA_NA_1        -0.155453547 -0.024188044
#> nu_y_alpha_NA_NA_2        -0.114936578 -0.001180303
#> nu_y_alpha_NA_NA_3         0.038323233  0.137730992
#> nu_y_alpha_NA_NA_4        -0.021057566  0.098196788
#> nu_y_alpha_NA_NA_5        -0.093881338  0.022357500
#> nu_y_alpha_NA_NA_6         0.061422770  0.188127508
#> nu_y_alpha_NA_NA_7        -0.014523680  0.109590119
#> nu_y_alpha_NA_NA_8        -0.147881702 -0.021442454
#> nu_y_alpha_NA_NA_9        -0.076286714  0.033948755
#> nu_y_alpha_NA_NA_10       -0.129269421 -0.013270978
#> nu_y_alpha_NA_NA_11        0.048517259  0.177117076
#> nu_y_alpha_NA_NA_12       -0.085159026  0.040449678
#> nu_y_alpha_NA_NA_13       -0.033303822  0.078936774
#> nu_y_alpha_NA_NA_14       -0.064016616  0.054087502
#> nu_y_alpha_NA_NA_15       -0.134813750 -0.016147313
#> nu_y_alpha_NA_NA_16        0.124877931  0.245423309
#> nu_y_alpha_NA_NA_17       -0.012199525  0.110257217
#> nu_y_alpha_NA_NA_18       -0.279136592 -0.151116917
#> nu_y_alpha_NA_NA_19       -0.018235391  0.105667617
#> nu_y_alpha_NA_NA_20       -0.037903493  0.080943615
#> nu_y_alpha_NA_NA_21       -0.192293953 -0.067132377
#> nu_y_alpha_NA_NA_22       -0.078905110  0.040456895
#> nu_y_alpha_NA_NA_23       -0.188551263 -0.061730625
#> nu_y_alpha_NA_NA_24       -0.156862230 -0.029591627
#> nu_y_alpha_NA_NA_25       -0.098932956  0.027841651
#> nu_y_alpha_NA_NA_26       -0.187374221 -0.058029975
#> nu_y_alpha_NA_NA_27        0.010014378  0.139698199
#> nu_y_alpha_NA_NA_28       -0.100250355  0.022370860
#> nu_y_alpha_NA_NA_29       -0.135928023 -0.017290691
#> nu_y_alpha_NA_NA_30        0.043455239  0.168606425
#> nu_y_alpha_NA_NA_31       -0.080108549  0.034451130
#> nu_y_alpha_NA_NA_32       -0.119678589  0.017649169
#> nu_y_alpha_NA_NA_33       -0.004042130  0.097744338
#> nu_y_alpha_NA_NA_34        0.031173945  0.149431552
#> nu_y_alpha_NA_NA_35       -0.014669270  0.096811319
#> nu_y_alpha_NA_NA_36        0.031743418  0.156677691
#> nu_y_alpha_NA_NA_37        0.014610762  0.127281884
#> nu_y_alpha_NA_NA_38       -0.060141294  0.052915017
#> nu_y_alpha_NA_NA_39       -0.059315742  0.067072075
#> nu_y_alpha_NA_NA_40       -0.098092118  0.021776301
#> nu_y_alpha_NA_NA_41       -0.133772224 -0.002735198
#> nu_y_alpha_NA_NA_42       -0.037481210  0.086298312
#> nu_y_alpha_NA_NA_43       -0.180734969 -0.046712628
#> nu_y_alpha_NA_NA_44        0.151348804  0.276526877
#> nu_y_alpha_NA_NA_45        0.033425966  0.151472920
#> nu_y_alpha_NA_NA_46       -0.140921132 -0.003743982
#> nu_y_alpha_NA_NA_47       -0.024131049  0.092747521
#> nu_y_alpha_NA_NA_48       -0.086021797  0.025018396
#> nu_y_alpha_NA_NA_49       -0.014691722  0.105828744
#> nu_y_alpha_NA_NA_50       -0.059368234  0.045942031