Credible Intervals for Dynamite Model Parameters

Extracts credible intervals from dynamitefit object.

Usage

# S3 method for dynamitefit
confint(object, parm, level = 0.95, ...)

Arguments

object

[dynamitefit]
The model fit object.

parm

Ignored.

level

[numeric(1)]
Credible interval width.

...

Ignored.

Value

The rows of the resulting matrix will be named using the following logic: {parameter}_{time}_{category}_{group} where parameter is the name of the parameter, time is the time index of the parameter, category specifies the level of the response the parameter is related to if the response is categorical, and group determines which group of observations the parameter is related to in the case of random effects and loadings. Non-applicable fields in the this syntax are set to NA.

See also

Examples

data.table::setDTthreads(1) # For CRAN
confint(gaussian_example_fit, level = 0.9)
#>                                     5%           95%
#> alpha_y_1_NA_NA                     NA            NA
#> alpha_y_2_NA_NA            0.009002244  1.056375e-01
#> alpha_y_3_NA_NA            0.028521871  1.613490e-01
#> alpha_y_4_NA_NA            0.110352533  2.228091e-01
#> alpha_y_5_NA_NA            0.196190932  3.223306e-01
#> alpha_y_6_NA_NA            0.244023065  3.582638e-01
#> alpha_y_7_NA_NA            0.270211412  3.859599e-01
#> alpha_y_8_NA_NA            0.361812829  4.725911e-01
#> alpha_y_9_NA_NA            0.387984729  5.245017e-01
#> alpha_y_10_NA_NA           0.334659338  4.782931e-01
#> alpha_y_11_NA_NA           0.339125127  4.688978e-01
#> alpha_y_12_NA_NA           0.372706929  5.053667e-01
#> alpha_y_13_NA_NA           0.421895578  5.517268e-01
#> alpha_y_14_NA_NA           0.442377391  5.676580e-01
#> alpha_y_15_NA_NA           0.432519547  5.628703e-01
#> alpha_y_16_NA_NA           0.423582091  5.340399e-01
#> alpha_y_17_NA_NA           0.408947521  5.280125e-01
#> alpha_y_18_NA_NA           0.400110887  5.134354e-01
#> alpha_y_19_NA_NA           0.366024245  4.951527e-01
#> alpha_y_20_NA_NA           0.325526766  4.615143e-01
#> alpha_y_21_NA_NA           0.268498051  3.995335e-01
#> alpha_y_22_NA_NA           0.153020010  2.944117e-01
#> alpha_y_23_NA_NA          -0.010020276  1.194010e-01
#> alpha_y_24_NA_NA          -0.076294799  5.441740e-02
#> alpha_y_25_NA_NA          -0.037301562  8.602467e-02
#> alpha_y_26_NA_NA          -0.007976526  1.096131e-01
#> alpha_y_27_NA_NA           0.001940210  1.305184e-01
#> alpha_y_28_NA_NA           0.026028761  1.540972e-01
#> alpha_y_29_NA_NA           0.165531755  2.896290e-01
#> alpha_y_30_NA_NA           0.654957454  8.006164e-01
#> beta_y_z_NA_NA_NA          1.950652141  1.988513e+00
#> delta_y_x_1_NA_NA                   NA            NA
#> delta_y_x_2_NA_NA         -0.207889269 -1.110911e-01
#> delta_y_x_3_NA_NA         -0.968760315 -8.551090e-01
#> delta_y_x_4_NA_NA         -0.927587267 -8.459252e-01
#> delta_y_x_5_NA_NA         -0.952494437 -8.855338e-01
#> delta_y_x_6_NA_NA         -1.100517439 -1.029873e+00
#> delta_y_x_7_NA_NA         -1.179482693 -1.105107e+00
#> delta_y_x_8_NA_NA         -1.028542165 -9.610408e-01
#> delta_y_x_9_NA_NA         -0.941010122 -8.639187e-01
#> delta_y_x_10_NA_NA        -1.049712859 -9.708599e-01
#> delta_y_x_11_NA_NA        -1.224154523 -1.170493e+00
#> delta_y_x_12_NA_NA        -1.402172999 -1.331471e+00
#> delta_y_x_13_NA_NA        -1.489219547 -1.422477e+00
#> delta_y_x_14_NA_NA        -1.486477307 -1.421423e+00
#> delta_y_x_15_NA_NA        -1.449747776 -1.378779e+00
#> delta_y_x_16_NA_NA        -1.451752641 -1.383490e+00
#> delta_y_x_17_NA_NA        -1.457486825 -1.374348e+00
#> delta_y_x_18_NA_NA        -1.399721793 -1.336782e+00
#> delta_y_x_19_NA_NA        -1.381265924 -1.318495e+00
#> delta_y_x_20_NA_NA        -1.391065471 -1.313991e+00
#> delta_y_x_21_NA_NA        -1.309151384 -1.240601e+00
#> delta_y_x_22_NA_NA        -1.255251802 -1.176625e+00
#> delta_y_x_23_NA_NA        -1.275720993 -1.200149e+00
#> delta_y_x_24_NA_NA        -1.239645854 -1.166628e+00
#> delta_y_x_25_NA_NA        -1.102300422 -1.033547e+00
#> delta_y_x_26_NA_NA        -0.930554709 -8.688729e-01
#> delta_y_x_27_NA_NA        -0.774334796 -7.000450e-01
#> delta_y_x_28_NA_NA        -0.673341014 -5.981539e-01
#> delta_y_x_29_NA_NA        -0.678573632 -5.660270e-01
#> delta_y_x_30_NA_NA        -0.259426820 -1.684200e-01
#> delta_y_y_lag1_1_NA_NA              NA            NA
#> delta_y_y_lag1_2_NA_NA     0.026816867  1.110380e-01
#> delta_y_y_lag1_3_NA_NA     0.126757650  1.999316e-01
#> delta_y_y_lag1_4_NA_NA     0.133938568  1.861013e-01
#> delta_y_y_lag1_5_NA_NA     0.135175175  1.858805e-01
#> delta_y_y_lag1_6_NA_NA     0.155068676  2.063058e-01
#> delta_y_y_lag1_7_NA_NA     0.188705124  2.366087e-01
#> delta_y_y_lag1_8_NA_NA     0.211988196  2.544995e-01
#> delta_y_y_lag1_9_NA_NA     0.256624348  3.055813e-01
#> delta_y_y_lag1_10_NA_NA    0.340138304  3.873894e-01
#> delta_y_y_lag1_11_NA_NA    0.435828713  4.779582e-01
#> delta_y_y_lag1_12_NA_NA    0.507383223  5.447754e-01
#> delta_y_y_lag1_13_NA_NA    0.513106923  5.468509e-01
#> delta_y_y_lag1_14_NA_NA    0.473156423  5.070559e-01
#> delta_y_y_lag1_15_NA_NA    0.417581475  4.502356e-01
#> delta_y_y_lag1_16_NA_NA    0.387865717  4.166519e-01
#> delta_y_y_lag1_17_NA_NA    0.375599875  4.102026e-01
#> delta_y_y_lag1_18_NA_NA    0.378165042  4.078493e-01
#> delta_y_y_lag1_19_NA_NA    0.367517851  3.999047e-01
#> delta_y_y_lag1_20_NA_NA    0.337702053  3.742921e-01
#> delta_y_y_lag1_21_NA_NA    0.307626956  3.432930e-01
#> delta_y_y_lag1_22_NA_NA    0.265536278  3.057002e-01
#> delta_y_y_lag1_23_NA_NA    0.232734374  2.753360e-01
#> delta_y_y_lag1_24_NA_NA    0.220570714  2.624469e-01
#> delta_y_y_lag1_25_NA_NA    0.214293304  2.588125e-01
#> delta_y_y_lag1_26_NA_NA    0.179395043  2.216431e-01
#> delta_y_y_lag1_27_NA_NA    0.109867983  1.591547e-01
#> delta_y_y_lag1_28_NA_NA    0.016004880  7.401490e-02
#> delta_y_y_lag1_29_NA_NA   -0.067305100 -3.625804e-03
#> delta_y_y_lag1_30_NA_NA    0.013058338  9.673072e-02
#> tau_y_x_NA_NA_NA           0.266567301  4.689455e-01
#> tau_y_y_lag1_NA_NA_NA      0.078033589  1.335140e-01
#> tau_alpha_y_NA_NA_NA       0.154095889  2.968183e-01
#> sigma_nu_y_alpha_NA_NA_NA  0.077116597  1.099316e-01
#> sigma_y_NA_NA_NA           0.191117875  2.031591e-01
#> nu_y_alpha_NA_NA_1        -0.151339354 -2.451919e-02
#> nu_y_alpha_NA_NA_2        -0.122352377  1.102223e-02
#> nu_y_alpha_NA_NA_3         0.038274846  1.481011e-01
#> nu_y_alpha_NA_NA_4        -0.013775195  1.109872e-01
#> nu_y_alpha_NA_NA_5        -0.095425858  1.716568e-02
#> nu_y_alpha_NA_NA_6         0.062038055  1.843678e-01
#> nu_y_alpha_NA_NA_7        -0.006864504  1.057019e-01
#> nu_y_alpha_NA_NA_8        -0.133879637 -1.769273e-02
#> nu_y_alpha_NA_NA_9        -0.073318600  4.296864e-02
#> nu_y_alpha_NA_NA_10       -0.139836882 -2.065564e-02
#> nu_y_alpha_NA_NA_11        0.056658360  1.843188e-01
#> nu_y_alpha_NA_NA_12       -0.084876721  2.699963e-02
#> nu_y_alpha_NA_NA_13       -0.042553295  8.548068e-02
#> nu_y_alpha_NA_NA_14       -0.060767408  6.270848e-02
#> nu_y_alpha_NA_NA_15       -0.137067235 -1.478445e-02
#> nu_y_alpha_NA_NA_16        0.115574047  2.472857e-01
#> nu_y_alpha_NA_NA_17       -0.004644597  1.143268e-01
#> nu_y_alpha_NA_NA_18       -0.283868947 -1.627500e-01
#> nu_y_alpha_NA_NA_19       -0.011074923  1.076906e-01
#> nu_y_alpha_NA_NA_20       -0.023714174  8.130681e-02
#> nu_y_alpha_NA_NA_21       -0.181201451 -6.432070e-02
#> nu_y_alpha_NA_NA_22       -0.072246351  4.094765e-02
#> nu_y_alpha_NA_NA_23       -0.186257320 -7.294741e-02
#> nu_y_alpha_NA_NA_24       -0.145979516 -2.773658e-02
#> nu_y_alpha_NA_NA_25       -0.102252394  7.285286e-03
#> nu_y_alpha_NA_NA_26       -0.180454723 -6.476092e-02
#> nu_y_alpha_NA_NA_27        0.014122254  1.412671e-01
#> nu_y_alpha_NA_NA_28       -0.093013006  1.332960e-02
#> nu_y_alpha_NA_NA_29       -0.141011503 -2.206665e-02
#> nu_y_alpha_NA_NA_30        0.047211339  1.682661e-01
#> nu_y_alpha_NA_NA_31       -0.087600750  4.563315e-02
#> nu_y_alpha_NA_NA_32       -0.099317768  1.908678e-02
#> nu_y_alpha_NA_NA_33       -0.006336926  1.075948e-01
#> nu_y_alpha_NA_NA_34        0.033127504  1.596061e-01
#> nu_y_alpha_NA_NA_35       -0.019156439  1.099636e-01
#> nu_y_alpha_NA_NA_36        0.032406102  1.521179e-01
#> nu_y_alpha_NA_NA_37        0.027709223  1.387856e-01
#> nu_y_alpha_NA_NA_38       -0.058053755  5.365393e-02
#> nu_y_alpha_NA_NA_39       -0.052725774  5.563560e-02
#> nu_y_alpha_NA_NA_40       -0.092522810  2.119350e-02
#> nu_y_alpha_NA_NA_41       -0.127034574 -5.147128e-05
#> nu_y_alpha_NA_NA_42       -0.032490083  7.958533e-02
#> nu_y_alpha_NA_NA_43       -0.180912191 -5.796359e-02
#> nu_y_alpha_NA_NA_44        0.157166662  2.704791e-01
#> nu_y_alpha_NA_NA_45        0.034851799  1.618409e-01
#> nu_y_alpha_NA_NA_46       -0.123297277 -1.116419e-05
#> nu_y_alpha_NA_NA_47       -0.026576678  7.932361e-02
#> nu_y_alpha_NA_NA_48       -0.081473021  1.922228e-02
#> nu_y_alpha_NA_NA_49       -0.008609759  1.152341e-01
#> nu_y_alpha_NA_NA_50       -0.053920539  6.273405e-02