Credible Intervals for dynamite Model Parameters

Extracts credible intervals from dynamitefit object.

Usage

# S3 method for class 'dynamitefit'
confint(object, parm, level = 0.95, ...)

Arguments

object

[dynamitefit]
The model fit object.

parm

Ignored.

level

[numeric(1)]
Credible interval width.

...

Ignored.

Value

The rows of the resulting matrix will be named using the following logic: {parameter}_{time}_{category}_{group} where parameter is the name of the parameter, time is the time index of the parameter, category specifies the level of the response the parameter is related to if the response is categorical, and group determines which group of observations the parameter is related to in the case of random effects and loadings. Non-applicable fields in the this syntax are set to NA.

See also

Examples

data.table::setDTthreads(1) # For CRAN
confint(gaussian_example_fit, level = 0.9)
#>                                      5%          95%
#> alpha_y_1_NA_NA                      NA           NA
#> alpha_y_2_NA_NA            4.523993e-03  0.102596029
#> alpha_y_3_NA_NA            3.359892e-02  0.160444793
#> alpha_y_4_NA_NA            1.027873e-01  0.232540083
#> alpha_y_5_NA_NA            1.990054e-01  0.325578658
#> alpha_y_6_NA_NA            2.495795e-01  0.372449363
#> alpha_y_7_NA_NA            2.757729e-01  0.399835032
#> alpha_y_8_NA_NA            3.646104e-01  0.480932301
#> alpha_y_9_NA_NA            3.919917e-01  0.519677560
#> alpha_y_10_NA_NA           3.500797e-01  0.491642343
#> alpha_y_11_NA_NA           3.403296e-01  0.472880570
#> alpha_y_12_NA_NA           3.721109e-01  0.517069255
#> alpha_y_13_NA_NA           4.256413e-01  0.563701653
#> alpha_y_14_NA_NA           4.508549e-01  0.566089018
#> alpha_y_15_NA_NA           4.244461e-01  0.563042309
#> alpha_y_16_NA_NA           4.056710e-01  0.539230818
#> alpha_y_17_NA_NA           4.032323e-01  0.533287337
#> alpha_y_18_NA_NA           3.952055e-01  0.526240097
#> alpha_y_19_NA_NA           3.632834e-01  0.493968029
#> alpha_y_20_NA_NA           3.202149e-01  0.452028381
#> alpha_y_21_NA_NA           2.602953e-01  0.397567475
#> alpha_y_22_NA_NA           1.545481e-01  0.288289473
#> alpha_y_23_NA_NA           8.637629e-03  0.118678163
#> alpha_y_24_NA_NA          -5.573237e-02  0.053260588
#> alpha_y_25_NA_NA          -3.262542e-02  0.078526519
#> alpha_y_26_NA_NA          -3.935294e-06  0.100660195
#> alpha_y_27_NA_NA           1.267877e-02  0.122305659
#> alpha_y_28_NA_NA           3.753272e-02  0.162030500
#> alpha_y_29_NA_NA           1.727556e-01  0.293282789
#> alpha_y_30_NA_NA           6.511299e-01  0.803205032
#> beta_y_z_NA_NA_NA          1.949136e+00  1.984239847
#> delta_y_x_1_NA_NA                    NA           NA
#> delta_y_x_2_NA_NA         -2.140651e-01 -0.109780116
#> delta_y_x_3_NA_NA         -9.661983e-01 -0.852173569
#> delta_y_x_4_NA_NA         -9.239329e-01 -0.839499886
#> delta_y_x_5_NA_NA         -9.587950e-01 -0.884174732
#> delta_y_x_6_NA_NA         -1.098449e+00 -1.033927809
#> delta_y_x_7_NA_NA         -1.175602e+00 -1.100404037
#> delta_y_x_8_NA_NA         -1.035984e+00 -0.962228004
#> delta_y_x_9_NA_NA         -9.418789e-01 -0.864142998
#> delta_y_x_10_NA_NA        -1.046748e+00 -0.974574117
#> delta_y_x_11_NA_NA        -1.223947e+00 -1.165802005
#> delta_y_x_12_NA_NA        -1.404571e+00 -1.332891359
#> delta_y_x_13_NA_NA        -1.486715e+00 -1.424051353
#> delta_y_x_14_NA_NA        -1.487714e+00 -1.422547170
#> delta_y_x_15_NA_NA        -1.446169e+00 -1.381804106
#> delta_y_x_16_NA_NA        -1.446872e+00 -1.389537774
#> delta_y_x_17_NA_NA        -1.456185e+00 -1.384385479
#> delta_y_x_18_NA_NA        -1.398059e+00 -1.335546923
#> delta_y_x_19_NA_NA        -1.381143e+00 -1.316983285
#> delta_y_x_20_NA_NA        -1.388983e+00 -1.313945369
#> delta_y_x_21_NA_NA        -1.306640e+00 -1.240737017
#> delta_y_x_22_NA_NA        -1.253168e+00 -1.176769426
#> delta_y_x_23_NA_NA        -1.279166e+00 -1.211951431
#> delta_y_x_24_NA_NA        -1.245856e+00 -1.170451671
#> delta_y_x_25_NA_NA        -1.109322e+00 -1.036566113
#> delta_y_x_26_NA_NA        -9.370633e-01 -0.864553806
#> delta_y_x_27_NA_NA        -7.805103e-01 -0.699271552
#> delta_y_x_28_NA_NA        -6.777650e-01 -0.597374303
#> delta_y_x_29_NA_NA        -6.759793e-01 -0.573684019
#> delta_y_x_30_NA_NA        -2.639551e-01 -0.170006895
#> delta_y_y_lag1_1_NA_NA               NA           NA
#> delta_y_y_lag1_2_NA_NA     2.693359e-02  0.111502050
#> delta_y_y_lag1_3_NA_NA     1.276624e-01  0.201562837
#> delta_y_y_lag1_4_NA_NA     1.332381e-01  0.185436403
#> delta_y_y_lag1_5_NA_NA     1.362361e-01  0.184274517
#> delta_y_y_lag1_6_NA_NA     1.577467e-01  0.201269787
#> delta_y_y_lag1_7_NA_NA     1.869231e-01  0.231708845
#> delta_y_y_lag1_8_NA_NA     2.108034e-01  0.251270852
#> delta_y_y_lag1_9_NA_NA     2.574953e-01  0.302521289
#> delta_y_y_lag1_10_NA_NA    3.359790e-01  0.384805049
#> delta_y_y_lag1_11_NA_NA    4.385428e-01  0.477644495
#> delta_y_y_lag1_12_NA_NA    5.064863e-01  0.543360361
#> delta_y_y_lag1_13_NA_NA    5.126052e-01  0.543766133
#> delta_y_y_lag1_14_NA_NA    4.751510e-01  0.503739288
#> delta_y_y_lag1_15_NA_NA    4.226538e-01  0.452675141
#> delta_y_y_lag1_16_NA_NA    3.899357e-01  0.420632883
#> delta_y_y_lag1_17_NA_NA    3.754850e-01  0.411599622
#> delta_y_y_lag1_18_NA_NA    3.777250e-01  0.409423065
#> delta_y_y_lag1_19_NA_NA    3.671966e-01  0.399022882
#> delta_y_y_lag1_20_NA_NA    3.396516e-01  0.377676931
#> delta_y_y_lag1_21_NA_NA    3.079593e-01  0.343735055
#> delta_y_y_lag1_22_NA_NA    2.680208e-01  0.305820595
#> delta_y_y_lag1_23_NA_NA    2.366441e-01  0.271021339
#> delta_y_y_lag1_24_NA_NA    2.183647e-01  0.256928862
#> delta_y_y_lag1_25_NA_NA    2.152175e-01  0.256754306
#> delta_y_y_lag1_26_NA_NA    1.808735e-01  0.220963619
#> delta_y_y_lag1_27_NA_NA    1.133814e-01  0.154441741
#> delta_y_y_lag1_28_NA_NA    2.137628e-02  0.072069728
#> delta_y_y_lag1_29_NA_NA   -6.947449e-02 -0.008900573
#> delta_y_y_lag1_30_NA_NA    9.616872e-03  0.098969316
#> nu_y_alpha_NA_NA_1        -1.489707e-01 -0.034425961
#> nu_y_alpha_NA_NA_2        -1.101465e-01  0.006639100
#> nu_y_alpha_NA_NA_3         2.966781e-02  0.153729578
#> nu_y_alpha_NA_NA_4        -2.249542e-02  0.092851192
#> nu_y_alpha_NA_NA_5        -1.041095e-01  0.021548818
#> nu_y_alpha_NA_NA_6         6.437379e-02  0.177625091
#> nu_y_alpha_NA_NA_7        -1.614390e-02  0.098079460
#> nu_y_alpha_NA_NA_8        -1.508397e-01 -0.022266947
#> nu_y_alpha_NA_NA_9        -8.749513e-02  0.028928615
#> nu_y_alpha_NA_NA_10       -1.412191e-01 -0.021662580
#> nu_y_alpha_NA_NA_11        5.074919e-02  0.175470198
#> nu_y_alpha_NA_NA_12       -8.573498e-02  0.037627269
#> nu_y_alpha_NA_NA_13       -3.651632e-02  0.093432080
#> nu_y_alpha_NA_NA_14       -6.406755e-02  0.051779504
#> nu_y_alpha_NA_NA_15       -1.334432e-01 -0.022891301
#> nu_y_alpha_NA_NA_16        1.189248e-01  0.247382674
#> nu_y_alpha_NA_NA_17       -4.465580e-03  0.107305388
#> nu_y_alpha_NA_NA_18       -2.810331e-01 -0.167231896
#> nu_y_alpha_NA_NA_19       -2.179973e-02  0.105130364
#> nu_y_alpha_NA_NA_20       -2.558402e-02  0.077005680
#> nu_y_alpha_NA_NA_21       -1.747152e-01 -0.059276472
#> nu_y_alpha_NA_NA_22       -7.679899e-02  0.050340768
#> nu_y_alpha_NA_NA_23       -1.877313e-01 -0.071091299
#> nu_y_alpha_NA_NA_24       -1.482093e-01 -0.033621879
#> nu_y_alpha_NA_NA_25       -1.007265e-01  0.017994418
#> nu_y_alpha_NA_NA_26       -1.800573e-01 -0.070422226
#> nu_y_alpha_NA_NA_27        2.113455e-02  0.139184008
#> nu_y_alpha_NA_NA_28       -1.012392e-01  0.017283459
#> nu_y_alpha_NA_NA_29       -1.425264e-01 -0.017553696
#> nu_y_alpha_NA_NA_30        4.230518e-02  0.164018405
#> nu_y_alpha_NA_NA_31       -9.036620e-02  0.034262600
#> nu_y_alpha_NA_NA_32       -1.037199e-01  0.030813037
#> nu_y_alpha_NA_NA_33       -1.829814e-03  0.110442403
#> nu_y_alpha_NA_NA_34        3.036013e-02  0.157468985
#> nu_y_alpha_NA_NA_35       -2.336594e-02  0.102503742
#> nu_y_alpha_NA_NA_36        3.356255e-02  0.169358423
#> nu_y_alpha_NA_NA_37        1.987934e-02  0.152544564
#> nu_y_alpha_NA_NA_38       -6.984231e-02  0.057836406
#> nu_y_alpha_NA_NA_39       -5.890937e-02  0.061691569
#> nu_y_alpha_NA_NA_40       -9.800524e-02  0.025874339
#> nu_y_alpha_NA_NA_41       -1.265697e-01 -0.001644795
#> nu_y_alpha_NA_NA_42       -3.286195e-02  0.088924471
#> nu_y_alpha_NA_NA_43       -1.672104e-01 -0.055155832
#> nu_y_alpha_NA_NA_44        1.552476e-01  0.279116344
#> nu_y_alpha_NA_NA_45        2.483351e-02  0.150806854
#> nu_y_alpha_NA_NA_46       -1.104029e-01  0.009186959
#> nu_y_alpha_NA_NA_47       -4.150920e-02  0.086011583
#> nu_y_alpha_NA_NA_48       -9.088547e-02  0.029514883
#> nu_y_alpha_NA_NA_49       -1.392437e-02  0.117751149
#> nu_y_alpha_NA_NA_50       -6.897436e-02  0.053294097
#> sigma_y_NA_NA_NA           1.920934e-01  0.205189135
#> sigma_nu_y_alpha_NA_NA_NA  7.913559e-02  0.119782251
#> tau_y_x_NA_NA_NA           2.533038e-01  0.497250450
#> tau_y_y_lag1_NA_NA_NA      7.639169e-02  0.139829649
#> tau_alpha_y_NA_NA_NA       1.386284e-01  0.283403259